+++ ChangeLog.md
@@ -13,6 +13,7 @@
     - Define new wantLanguageSwitch() RunContext method, thus bumping minor candidate version number
     - RunContext instance is now a singleton
     - Introduce new IAppI18n interface and iAppI18nWantSwitch() method replacing wantLanguageSwitch()
+    - Extend pwix:env-settings with environmentSettings() async function
 
 ### 1.0.0